move validation and shield to seperate file
